+++ docs/README-zh.md
@@ -1,7 +1,7 @@
 
 # AIAPA
 
-这是一个亚马逊产品分析工具，使用Gemini驱动，目前还没有完成
+这是一个亚马逊产品分析工具，使用Gemini驱动，目前还没有完成！
 
 已经可以通过get指令下载商品数据
 
@@ -33,7 +33,7 @@ aiapa get -q laptop -t 20 -mc 10 -r 5 -o "./output"
 
 ### 通过代码
 
-由于该包是基于ESM的，所以包无法在CJS上运行！也许我以后会想办法吧（以后）
+由于该包是基于ESM的，所以包无法在CJS上运行！也许我以后会想办法吧（以后:P
 
 从aiapa获取导入，并且运行任务！
 
@@ -67,6 +67,6 @@ app.run({
 
 ## 许可
 
-**不要用来做任何违法的事情！后果自负！**
+**不要用来做任何违法的事情！网络抓取实践应始终符合目标网站的服务条款和法律考虑**
 
 > 使用MIT许可证
