+++ src/commands/get.js
@@ -139,6 +139,10 @@ const ReviewSelectors = {
     },
 };
 
+export function registerDetailSelector(key, selector) {
+    Details[key] = selector;
+}
+
 /**
  * @typedef {Object} Review
  * @property {string} title
@@ -244,6 +248,10 @@ export async function search({ app, browser, page, search }) {
         if (!res || !res.length) throw new Error("No query provided, please provide by --query <string>");
         app.config.query = search = res;
     }
+    if(app.config.lowRam && app.config.maxConcurrency > 3) {
+        app.Logger.warn("Max Concurrecy is set to 3 because of low ram mode");
+        app.config.maxConcurrency = 5;
+    }
     let url = new URL(AMAZON_SEARCH_URL);
     url.searchParams.append("k", search);
     url.searchParams.append("s", "exact-aware-popularity-rank");
@@ -276,7 +284,7 @@ export async function search({ app, browser, page, search }) {
     let bar = createProgressBar();
     bar.start(app.config.maxTask, 0);
 
-    let products = [], pool = new TaskPool(app.config.maxConcurrency, 1).addTasks(
+    let products = [], pool = new TaskPool(app.config.maxConcurrency, app.config.lowRam ? 3 * 1000 : 0).addTasks(
         links.slice(0, Number(app.config.maxTask)).map((link) => async () => {
             return await browser.page(async (page) => {
                 await browser.blockResources(page, config.blockedResourceTypes);
@@ -322,7 +330,7 @@ async function getDetails({ app, browser, page }) {
                 value: await browser.select(page, Details[key])
             };
         } catch (err) {
-            app.Logger.warn("Failed to get details: " + key);
+            app.Logger.warn("Failed to get details: " + key + " on page: " + page.url());
             app.Logger.warn(err);
             return {
                 key,
